pyproject.toml
[project]
name = "citylogix-ai-inference"
version = "0.1.0"
description = "Run segmentation inference on road/pavement images using multiple AI model architectures"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.44",
    "pillow>=10.0",
    "opencv-python>=4.8",
    "pycocotools>=2.0",
    "typer>=0.12",
    "pydantic>=2.0",
    "loguru>=0.7",
    "numpy>=1.24",
    "tqdm>=4.66",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.1",
    "ipykernel>=6.0",
]

onnx = [
    "onnxruntime>=1.16",
]

onnx-gpu = [
    "onnxruntime-gpu>=1.16",
]

fiftyone = [
    "fiftyone>=0.23",
]

[project.scripts]
citylogix-infer = "citylogix_ai_inference.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/citylogix_ai_inference"]

[tool.ruff]
line-length = 100
target-version = "py311"
